Podcast Condenser-20250106T081355Z-001/Podcast Condenser/requirements.txt
openai~=0.28.0
Flask~=3.1.0
Flask-Cors~=5.0.0
python-dotenv~=1.0.1
youtube-transcript-api~=0.6.3